+++ sdk/resourcemanager/storageactions/armstorageactions/models_serde.go
@@ -366,6 +366,37 @@ func (s *StorageTaskAssignmentsListResult) UnmarshalJSON(data []byte) error {
 	return nil
 }
 
+// MarshalJSON implements the json.Marshaller interface for type StorageTaskListResult.
+func (s StorageTaskListResult) MarshalJSON() ([]byte, error) {
+	objectMap := make(map[string]any)
+	populate(objectMap, "nextLink", s.NextLink)
+	populate(objectMap, "value", s.Value)
+	return json.Marshal(objectMap)
+}
+
+// UnmarshalJSON implements the json.Unmarshaller interface for type StorageTaskListResult.
+func (s *StorageTaskListResult) UnmarshalJSON(data []byte) error {
+	var rawMsg map[string]json.RawMessage
+	if err := json.Unmarshal(data, &rawMsg); err != nil {
+		return fmt.Errorf("unmarshalling type %T: %v", s, err)
+	}
+	for key, val := range rawMsg {
+		var err error
+		switch key {
+		case "nextLink":
+			err = unpopulate(val, "NextLink", &s.NextLink)
+			delete(rawMsg, key)
+		case "value":
+			err = unpopulate(val, "Value", &s.Value)
+			delete(rawMsg, key)
+		}
+		if err != nil {
+			return fmt.Errorf("unmarshalling type %T: %v", s, err)
+		}
+	}
+	return nil
+}
+
 // MarshalJSON implements the json.Marshaller interface for type StorageTaskOperation.
 func (s StorageTaskOperation) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]any)
@@ -865,16 +896,20 @@ func (s *StorageTaskUpdateParameters) UnmarshalJSON(data []byte) error {
 	return nil
 }
 
-// MarshalJSON implements the json.Marshaller interface for type StorageTasksListResult.
-func (s StorageTasksListResult) MarshalJSON() ([]byte, error) {
+// MarshalJSON implements the json.Marshaller interface for type StorageTaskUpdateProperties.
+func (s StorageTaskUpdateProperties) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]any)
-	populate(objectMap, "nextLink", s.NextLink)
-	populate(objectMap, "value", s.Value)
+	populate(objectMap, "action", s.Action)
+	populateDateTimeRFC3339(objectMap, "creationTimeInUtc", s.CreationTimeInUTC)
+	populate(objectMap, "description", s.Description)
+	populate(objectMap, "enabled", s.Enabled)
+	populate(objectMap, "provisioningState", s.ProvisioningState)
+	populate(objectMap, "taskVersion", s.TaskVersion)
 	return json.Marshal(objectMap)
 }
 
-// UnmarshalJSON implements the json.Unmarshaller interface for type StorageTasksListResult.
-func (s *StorageTasksListResult) UnmarshalJSON(data []byte) error {
+// UnmarshalJSON implements the json.Unmarshaller interface for type StorageTaskUpdateProperties.
+func (s *StorageTaskUpdateProperties) UnmarshalJSON(data []byte) error {
 	var rawMsg map[string]json.RawMessage
 	if err := json.Unmarshal(data, &rawMsg); err != nil {
 		return fmt.Errorf("unmarshalling type %T: %v", s, err)
@@ -882,11 +917,23 @@ func (s *StorageTasksListResult) UnmarshalJSON(data []byte) error {
 	for key, val := range rawMsg {
 		var err error
 		switch key {
-		case "nextLink":
-			err = unpopulate(val, "NextLink", &s.NextLink)
+		case "action":
+			err = unpopulate(val, "Action", &s.Action)
 			delete(rawMsg, key)
-		case "value":
-			err = unpopulate(val, "Value", &s.Value)
+		case "creationTimeInUtc":
+			err = unpopulateDateTimeRFC3339(val, "CreationTimeInUTC", &s.CreationTimeInUTC)
+			delete(rawMsg, key)
+		case "description":
+			err = unpopulate(val, "Description", &s.Description)
+			delete(rawMsg, key)
+		case "enabled":
+			err = unpopulate(val, "Enabled", &s.Enabled)
+			delete(rawMsg, key)
+		case "provisioningState":
+			err = unpopulate(val, "ProvisioningState", &s.ProvisioningState)
+			delete(rawMsg, key)
+		case "taskVersion":
+			err = unpopulate(val, "TaskVersion", &s.TaskVersion)
 			delete(rawMsg, key)
 		}
 		if err != nil {